Fix detecting app to open url for chat
Should fix #160.

There is a builtin command in Neovim (I think from 0.10 up) for opening
a url in the dedicated app for the OS.
So, there shouldn't need to be a check for the OS, we can just pass it
in to the open function.

This also solves the issue that doing it with the OS checking means WSL
tries with xdg-open and fails.